try hull as its own library
