+++ .vscode/settings.json
@@ -121,7 +121,7 @@
     "-DBUILD_SHARED_LIBS=ON",
     "-DMANIFOLD_HULL=ON",
     "-DMANIFOLD_CBIND=ON",
-    "-DMANIFOLD_PYBIND=ON",
+    "-DMANIFOLD_PYBIND=OFF",
     "-DMANIFOLD_EXPORT=ON",
     "-DMANIFOLD_DEBUG=ON",
     "-DMANIFOLD_PAR=TBB",
