+++ bindings/python/CMakeLists.txt
@@ -20,7 +20,7 @@ nanobind_add_module(
   NB_STATIC STABLE_ABI LTO
   autogen_docstrings.inl
   manifold3d.cpp)
-target_link_libraries(manifold3d PRIVATE manifold sdf polygon)
+target_link_libraries(manifold3d PRIVATE manifold sdf polygon hull)
 target_compile_options(manifold3d PRIVATE ${MANIFOLD_FLAGS} -DMODULE_NAME=manifold3d)
 target_compile_features(manifold3d PUBLIC cxx_std_17)
 set_target_properties(manifold3d PROPERTIES OUTPUT_NAME "manifold3d")
@@ -31,6 +31,7 @@ set(DOCSTRING_DEPS
   ${CMAKE_CURRENT_SOURCE_DIR}/../../src/manifold/src/manifold.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/../../src/manifold/src/constructors.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/../../src/manifold/src/sort.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/../../src/manifold/module/hull.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/../../src/cross_section/src/cross_section.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/../../src/polygon/src/polygon.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/../../src/utilities/include/public.h
