+++ src/manifold/CMakeLists.txt
@@ -14,7 +14,7 @@
 
 project (manifold)
 
-file(GLOB_RECURSE SOURCE_FILES CONFIGURE_DEPENDS *.cpp)
+file(GLOB_RECURSE SOURCE_FILES CONFIGURE_DEPENDS src/*.cpp)
 add_library(${PROJECT_NAME} ${SOURCE_FILES})
 
 target_include_directories(${PROJECT_NAME} PUBLIC
@@ -24,11 +24,6 @@ target_link_libraries(${PROJECT_NAME}
     PUBLIC utilities cross_section
     PRIVATE collider polygon
 )
-if(MANIFOLD_HULL)
-target_link_libraries(${PROJECT_NAME}
-    PRIVATE quickhull
-)
-endif()
 
 target_compile_options(${PROJECT_NAME} PRIVATE ${MANIFOLD_FLAGS})
 
@@ -37,4 +32,17 @@ target_compile_features(${PROJECT_NAME}
 )
 
 install(TARGETS ${PROJECT_NAME} EXPORT manifoldTargets)
+
+if(MANIFOLD_HULL)
+add_library(hull module/hull.cpp)
+target_link_libraries(hull
+    PRIVATE manifold quickhull
+)
+target_compile_options(hull PRIVATE ${MANIFOLD_FLAGS})
+target_compile_features(hull
+    PUBLIC cxx_std_17
+)
+install(TARGETS hull EXPORT manifoldTargets)
+endif()
+
 install(FILES include/manifold.h DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/${CMAKE_PROJECT_NAME})
